\name{collection}
\alias{collection}
\alias{is.collection}
\alias{as.collection}
\alias{[.collection}
\alias{[<-.collection}
\title{collection objects}
\description{Deprecated, use LIST instead.}
\usage{
collection (..., resolve=TRUE, call=sys.call () )
is.collection (obj)
as.collection (obj)
\method{[}{collection}(obj, \dots)
\method{[}{collection}(obj, \dots)<-value
}
\arguments{
\item{resolve}{.}
\item{call}{.}
\item{obj}{.}
\item{value}{.}
\item{\dots}{.}
}

