\name{FUNCTION}
\alias{FUNCTION}
\alias{is.FUNCTION}
\alias{clone.FUNCTION}
\alias{extendf}
\alias{print.FUNCTION}
\alias{$.FUNCTION}
\alias{$<-.FUNCTION}
\title{Enhanced Functions}
\description{Enhanced function objects, refer to the relevant package vignette for more information.}
\usage{
FUNCTION (f, \dots, call)
is.FUNCTION (f)
extendf (f, subclass, g, \dots, attributes)
\method{print}{FUNCTION}(f, \dots)
\method{clone}{FUNCTION}(f, \dots)
\method{$}{FUNCTION}(f, name)
\method{$}{FUNCTION}(f, name) <- value
}
\arguments{
\item{f}{A function.}
\item{call}{The call used to resolve argument names.}
\item{subclass}{.}
\item{g}{.}
\item{attributes}{.}
\item{name}{.}
\item{value}{.}
\item{\dots}{.}
}


