\name{masks}
\alias{print}
\alias{format}
\alias{summary}
\alias{plot}
\alias{lines}
\alias{points}
\alias{fitted}
\alias{residuals}
\alias{as.list}
\alias{as.data.frame}
\alias{mean}
\title{mask functions for standard generics}
\description{These functions mask standard generics, for the purposes of getting around R check's S3 restrictions.}
\usage{
print (\dots)
format (\dots)
summary (\dots)
plot (\dots)
lines (\dots)
points (\dots)
fitted (\dots)
residuals (\dots)
as.list (\dots)
as.data.frame (\dots)
mean (\dots)
}
\arguments{
\item{\dots}{The arguments.}
}




