\name{misc}
\alias{fit}
\alias{ifst}
\alias{if0}
\alias{if1}
\alias{iter}
\alias{itobj}
\title{some experimental functions}
\description{These functions are being reviewed, and may be removed.}
\usage{
fit (\dots)
ifst (obj)
if0 (obj)
if1 (obj)
iter (n, min=1)
itobj (obj)
}
\arguments{
\item{obj}{.}
\item{n}{.}
\item{min}{.}
\item{\dots}{.}
}



