\name{preview}
\alias{preview}
\alias{preview.TABLE}
\alias{preview.MATRIX}
\alias{preview.VECTOR}
\alias{preview.data.frame}
\alias{preview.matrix}
\alias{preview.default}
\title{preview object}
\description{Preview functions.}
\usage{
preview (\dots)
\method{preview}{TABLE}(x, n=5, \dots)
\method{preview}{MATRIX}(x, n=5, \dots)
\method{preview}{VECTOR}(x, n=5, \dots)
\method{preview}{data.frame}(x, n=5, \dots)
\method{preview}{matrix}(x, n=5, \dots)
\method{preview}{default}(x, \dots)
}
\arguments{
\item{x}{.}
\item{n}{.}
\item{\dots}{.}
}


