\name{SERIES}
\alias{SERIES}
\alias{print.SERIES}
\alias{plot.SERIES}
\alias{lines.SERIES}
\alias{points.SERIES}
\alias{series1}
\title{series-based functions}
\description{The SERIES constructor creates a function, represented by a series of points. Mostly limited to regularly spaced series with linear interpolation. Don't call series.1 directly.}
\usage{
SERIES (x, y, regular=TRUE, f=series1)
\method{print}{SERIES}(f, \dots)
\method{plot}{SERIES}(f, \dots)
\method{lines}{SERIES}(f, \dots)
\method{points}{SERIES}(f, \dots)
series1 (u)
}
\arguments{
\item{x}{.}
\item{y}{.}
\item{regular}{.}
\item{f}{.}
\item{u}{.}
\item{\dots}{.}
}



