\name{SMOOTH}
\alias{SMOOTH}
\alias{print.SMOOTH}
\title{smooth functions}
\description{The SMOOTH constructor takes a raw series of points and return a SERIES function.}
\usage{
SMOOTH (x, y, n=10, degree=2, wf=.wf.std, smoothness, u)
\method{print}{SMOOTH}(f, \dots)
}
\arguments{
\item{x}{.}
\item{y}{.}
\item{n}{.}
\item{degree}{.}
\item{wf}{.}
\item{smoothness}{.}
\item{u}{.}
\item{f}{.}
\item{\dots}{.}
}

