\name{symbolic}
\alias{evalt}
\alias{definef}
\title{symbolic functions}
\description{The evalt function evaluates an expression, given as a character scalar. The definef function changes the body of a function, given as a character vector.}
\usage{
evalt (text)
definef (f, def, \dots)
}
\arguments{
\item{text}{.}
\item{f}{.}
\item{def}{.}
\item{\dots}{.}
}

