\name{TABLE}
\alias{TABLE}
\alias{is.TABLE}
\alias{dim.TABLE}
\alias{[.TABLE}
\alias{print.TABLE}
\alias{summary.TABLE}
\title{enhanced table objects}
\description{These functions are experimental.}
\usage{
TABLE (\dots, call=sys.call () )
is.TABLE (d)
\method{dim}{TABLE}(x)
\method{[}{TABLE}(obj, i, j, \dots)
\method{print}{TABLE}(d, \dots)
\method{summary}{TABLE}(d, \dots)
}
\arguments{
\item{call}{.}
\item{d}{.}
\item{x}{.}
\item{obj}{.}
\item{i}{.}
\item{j}{.}
\item{\dots}{.}
}


