\name{LIST}
\alias{LIST}
\alias{is.LIST}
\title{enhanced list objects}
\description{A LIST object extends a list object. By default it automatically resolves argument names.}
\usage{
LIST (\dots, call=sys.call () )
is.LIST (object)
}
\arguments{
\item{call}{The call containing the original arguments, that is, the call used to resolve the argument names.}
\item{object}{.}
\item{\dots}{Element objects.}
}

